Move actual logging in a function under utils.
